requirements.txt
annotated-types==0.5.0
anyio==3.7.1
boto3==1.28.40
botocore==1.31.40
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.103.1
h11==0.14.0
httptools==0.6.0
idna==3.4
jmespath==1.0.1
mangum==0.17.0
motor==3.3.1
passlib==1.7.4
pyasn1==0.5.0
pydantic==2.3.0
pydantic-core==2.6.3
pymongo==4.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
pyyaml==6.0.1
requests==2.31.0
rsa==4.9
s3transfer==0.6.2
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
stripe==6.2.0
typing-extensions==4.7.1
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3